+++ config.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Dolibase
+ * 
+ * Open source framework for Dolibarr ERP/CRM
+ *
+ * Copyright (c) 2018 - 2019
+ *
+ *
+ * @package     Dolibase
+ * @author      AXeL
+ * @copyright	Copyright (c) 2018 - 2019, AXeL-dev
+ * @license
+ * @link
+ * 
+ */
+
+global $dolibase_config;
+
+/**
+ * Dolibase main configuration
+ */
+
+$dolibase_config['main'] = array(
+	'version'             => '2.2.0-beta', // possible values: a.b.c-alpha, a.b.c-beta, a.b.c-rcX or a.b.c
+	'link'                => 'https://github.com/AXeL-dev/dolibase',
+	'path'                => get_dolibase_path(),
+	'tables'              => array('logs'),
+	'use_ajax_on_confirm' => 1 // value should be 1/0 instead of true/false
+);