+++ core/class/form_page.php
@@ -69,9 +69,9 @@ class FormPage extends Page
 	 */
 	public function askForConfirmation($url, $title, $question, $action, $question_param = '')
 	{
-		global $langs;
+		global $langs, $dolibase_config;
 
-		$this->body = $this->form->formconfirm($url, $langs->trans($title), $langs->trans($question, $question_param), $action, '', '', DOLIBASE_USE_AJAX_ON_CONFIRM);
+		$this->body = $this->form->formconfirm($url, $langs->trans($title), $langs->trans($question, $question_param), $action, '', '', $dolibase_config['main']['use_ajax_on_confirm']);
 	}
 
 	/**