+++ test/books/config.php
@@ -7,22 +7,24 @@ global $dolibase_config;
  */
 
 $dolibase_config['module'] = array(
-	'name'         => 'Books',
-	'desc'         => 'Books Managment Module',
-	'version'      => '1.5.0', // possible values: 'development', 'experimental', 'dolibarr' or version
-	'number'       => '450000', // e.: 110000 (avoid small numbers because they are used for core modules)
-	'family'       => 'Dolibase modules', // possible values: 'crm', 'financial', 'hr', 'projects', 'products', 'ecm', 'technic', 'other'
-	'position'     => 500,
-	'rights_class' => 'books', // key to reference module (for permissions, menus, etc.)
-	'url'          => '#', // e.: 'https://www.dolistore.com/my_module'
-	'folder'       => 'books', // advice: never use underscores in module folder name to avoid many problems especially with top menu icon, module widgets etc..
-	'picture'      => 'books.png', // your picture file should have 'object_' prefix to work
-	'dirs'         => array(), // directories to create when activating the module, e.: array('/modulefoldername/temp')
-	'dolibarr_min' => array(3, 8), // minimum Dolibarr version, e.: array(6, 0)
-	'php_min'      => array(5, 0), // minimum PHP version, e.: array(4, 0)
-	'depends'      => array(), // modules to activate when this one is activated, e.: array('modProduct')
-	'required_by'  => array(), // modules to disable if this one is disabled, e.: array('modFacture')
-	'conflit_with' => array() // modules who are in conflict with this module, e.: array('modProduct')
+	'name'          => 'Books',
+	'desc'          => 'Books Managment Module',
+	'version'       => '1.5.0', // possible values: 'development', 'experimental', 'dolibarr' or version
+	'number'        => '450000', // e.: 110000 (avoid small numbers because they are used for core modules)
+	'family'        => 'Dolibase modules', // possible values: 'crm', 'financial', 'hr', 'projects', 'products', 'ecm', 'technic', 'other'
+	'position'      => 500,
+	'rights_class'  => 'books', // key to reference module (for permissions, menus, etc.)
+	'url'           => '#', // e.: 'https://www.dolistore.com/my_module'
+	'folder'        => 'books', // advice: never use underscores in module folder name to avoid many problems especially with top menu icon, module widgets etc..
+	'picture'       => 'books.png', // your picture file should have 'object_' prefix to work
+	'dirs'          => array(), // directories to create when activating the module, e.: array('/modulefoldername/temp')
+	'dolibarr_min'  => array(3, 8), // minimum Dolibarr version, e.: array(6, 0)
+	'php_min'       => array(5, 0), // minimum PHP version, e.: array(4, 0)
+	'depends'       => array(), // modules to activate when this one is activated, e.: array('modProduct')
+	'required_by'   => array(), // modules to disable if this one is disabled, e.: array('modFacture')
+	'conflit_with'  => array(), // modules who are in conflict with this module, e.: array('modProduct')
+	'check_updates' => true, // check for module updates (module url should be filled also)
+	'enable_logs'   => true // enable saving logs when create/modify or delete an object
 );
 
 /**